Use uint8_t for extend instead of int
Since extend is between 0 and 100 and we can accommodate
ttl(uint32_t) and extend[4] into uint64

+++ include/nuster/cache.h
@@ -94,7 +94,7 @@ struct nst_cache_entry {
 
     /* For entries loaded from disk */
     uint32_t                ttl;
-    int                     extend[4];
+    uint8_t                 extend[4];
 
     /* see rule.extend */
     uint64_t                access[4];
